+++ CHANGELOG.md
@@ -5,6 +5,15 @@
 - MIG-825 - [BREAKING] Added parameter `runOptions` to `createMigrationRun` of `Migration/Run/RunService.php`
 - MIG-825 - Added the new final class `Migration/Run/RunOptions.php` to handle the options of a migration run
 - MIG-825 - [BREAKING] Added parameter `where` to `fetchIdentifiers` of `Profile/Shopware/Gateway/Local/Reader/AbstractReader.php`
+- MIG-825 - [BREAKING] Changed functions of `Profile/Shopware/Gateway/Local/Reader/AbstractReader.php` to be final:
+  - `setConnection`
+  - `addTableSelection`
+  - `buildArrayFromChunks`
+  - `cleanupResultSet`
+  - `fetchIdentifiers`
+  - `getDefaultShopLocale`
+  - `mapData`
+  - `getDataSetEntity`
 - MIG-838 - Add the meta information fields to the migration of category translations
 - MIG-839 - Add custom fields to the migration of category translations
 - MIG-899 - Changed behavior of the migration of seo urls. It now considers the URL case setting of shopware 5